Move study card filtering to deck filter feature
