--- src/entities/study-progress/@x/deck.ts
@@ -1,1 +1,0 @@
-export { createStudyProgressFromCard, isStudyProgressEligible } from "../model/rules";
